requirements.txt
# Core ML/NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# vLLM and LangGraph
vllm>=0.2.0
langgraph>=0.0.40
langgraph-checkpoint-sqlite>=0.0.1

# Data processing
datasets>=2.14.0
jsonlines>=3.1.0
numpy>=1.24.0
scipy>=1.11.0

# NLP utilities
nltk>=3.8.0

# API clients
openai>=1.0.0
requests>=2.31.0

# JSON parsing
json5>=0.9.14

# Progress bars
tqdm>=4.65.0

# System monitoring
psutil>=5.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Optional: GPU support for FAISS (uncomment if needed)
# faiss-gpu>=1.7.4

